Move redis check to bash script

+++ process_all.sh
@@ -26,9 +26,15 @@ for channel in $(jq -r '.[] | .channel' users.json); do
   echo "riot_ids: $riot_ids"
   for video in $(twitch-dl videos "$channel" --all -t archive -j | jq '.videos' | jq -r '.[].id'); do
       echo "Processing $video";
+      value_exists=$(redis-cli SISMEMBER processed "$video")
+      if [ "$value_exists" -eq 1 ]; then
+        echo "Video $video already processed. Skipping..."
+        continue
+      fi
       docker-compose run --rm --entrypoint "clip-cutter -v $video -r '$riot_ids' --remove-matches" clip_cutter || break;
       rclone move -P clips/ Nextcloud:ClipCutter/"$channel"/ || break;
       clean
+      redis-cli SADD processed "$video"
   done
 done
